requirements.txt
Flask
python-dotenv
openai
PyPDF2
pandas
pytesseract
Pillow
